remove wallet connect
