+++ package.json
@@ -31,11 +31,6 @@
         "@types/dotenv": "^8.2.0",
         "@types/react-native-vector-icons": "^6.4.4",
         "@types/secp256k1": "^3.5.2",
-        "@walletconnect/browser": "^1.0.0-beta.47",
-        "@walletconnect/core": "^1.0.0-beta.47",
-        "@walletconnect/react-native": "^1.0.0-beta.47",
-        "@walletconnect/types": "^1.0.0-beta.47",
-        "@walletconnect/utils": "^1.0.0-beta.47",
         "@zilliqa-js/account": "^0.8.4",
         "@zilliqa-js/crypto": "^0.8.1-alpha.0",
         "@zilliqa-js/util": "^0.8.3",
