+++ src/core/connect-extension/utils.ts
@@ -1,3 +1,7 @@
+import { getBlockchain } from '../blockchain/blockchain-factory';
+import { NavigationParams } from 'react-navigation';
+import { NavigationService } from '../../navigation/navigation-service';
+
 export const getUrlParams = (search = ``) => {
     const hashes = search.slice(search.indexOf('?') + 1).split('&');
     return hashes.reduce((params, hash) => {
@@ -12,4 +16,22 @@ export const getUrlParams = (search = ``) => {
 
         return { ...params, [key]: decodeURIComponent(val) };
     }, {});
+};
+
+export const signExtensionTransaction = payload => {
+    const data = payload.params[0];
+    const blockchainConfig = getBlockchain(data.selectedAccount.blockchain).config;
+    const navigationParams: NavigationParams = {
+        blockchain: data.selectedAccount.blockchain,
+        accountIndex: data.selectedAccount.index,
+        token: data.token,
+        tokenLogo: blockchainConfig.tokens[blockchainConfig.coin].icon,
+        extensionTransactionPayload: payload
+    };
+
+    if (NavigationService.getCurrentRoute() === 'Token') {
+        NavigationService.replace('Token', navigationParams);
+    } else {
+        NavigationService.navigate('Token', navigationParams);
+    }
 };
