--- src/core/wallet-connect/wallet-connect-web.ts
@@ -1,36 +1,0 @@
-export const WalletConnectWeb = (() => {
-    const connect = () => {
-        return Promise.resolve();
-    };
-
-    const subscribe = (eventName, callback) => {
-        return () => {
-            // @ts-ignore
-        };
-    };
-
-    const isConnected = () => true;
-    const setStore = store => true;
-
-    const disconnect = () => {
-        // @ts-ignore
-    };
-
-    const getState = () => {
-        return Promise.resolve();
-    };
-
-    const signTransaction = transaction => {
-        return Promise.resolve();
-    };
-
-    return {
-        connect,
-        disconnect,
-        subscribe,
-        isConnected,
-        setStore,
-        getState,
-        signTransaction
-    };
-})();
